buildathon-final/backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
chromadb
langchain
langchain-groq
langchain-community
langgraph
pypdf
httpx
requests
python-multipart
scikit-learn
firebase-admin
